Reorganize functions from secrets files in the agent package
